Fixed WeaponRegistry skill activation issues in multiplayer
Client WeaponRegistry instances were not kept in sync with the server, resulting in some skills not being able to be activated due to the held item not returning true from isSword or isWeapon.

The entire WeaponRegistry is now sent to each client as they log in, and changes made after that point (e.g. by Command) are kept in sync individually.